# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de GreenCupones App y movil",
        "description": "Documentación de la API para gestionar productos y ventas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/"
        }
    ],
    "paths": {
        "/attentions": {
            "get": {
                "tags": [
                    "Atención"
                ],
                "summary": "Listar todas las atenciones",
                "operationId": "a0204bd0f478d1f1804a652f27e8880d",
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Atención"
                ],
                "summary": "Crear una atención",
                "operationId": "3396e70c9c64fb82e2a85e7f77b1ebde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "appointment_date",
                                    "shoe_size",
                                    "footstep_type_left",
                                    "footstep_type_right",
                                    "foot_type_left",
                                    "foot_type_right",
                                    "heel_type_left",
                                    "heel_type_right"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-04"
                                    },
                                    "shoe_size": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 42
                                    },
                                    "footstep_type_left": {
                                        "type": "string",
                                        "example": "Pronador"
                                    },
                                    "footstep_type_right": {
                                        "type": "string",
                                        "example": "Supinador"
                                    },
                                    "foot_type_left": {
                                        "type": "string",
                                        "example": "Plano"
                                    },
                                    "foot_type_right": {
                                        "type": "string",
                                        "example": "Normal"
                                    },
                                    "heel_type_left": {
                                        "type": "string",
                                        "example": "Neutro"
                                    },
                                    "heel_type_right": {
                                        "type": "string",
                                        "example": "Pronador"
                                    },
                                    "observations": {
                                        "type": "string",
                                        "example": "Paciente presenta leve desviación..."
                                    },
                                    "back_standing_up_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "back_45_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "back_toes_up_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "from_chaplin_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "from_chaplin_toes_up_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "with_insoles_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Atención creada exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attentions/{id}": {
            "get": {
                "tags": [
                    "Atención"
                ],
                "summary": "Obtener una atención específica",
                "operationId": "785d3dfb4d301e0d404049b8bc602d55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la atención",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la atención"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Atención"
                ],
                "summary": "Actualizar una atención",
                "operationId": "c71ec9b202ae6a5e3dbae8d346ec367d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la atención a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-04"
                                    },
                                    "shoe_size": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 42
                                    },
                                    "footstep_type_left": {
                                        "type": "string",
                                        "example": "Pronador"
                                    },
                                    "footstep_type_right": {
                                        "type": "string",
                                        "example": "Supinador"
                                    },
                                    "foot_type_left": {
                                        "type": "string",
                                        "example": "Plano"
                                    },
                                    "foot_type_right": {
                                        "type": "string",
                                        "example": "Normal"
                                    },
                                    "heel_type_left": {
                                        "type": "string",
                                        "example": "Neutro"
                                    },
                                    "heel_type_right": {
                                        "type": "string",
                                        "example": "Pronador"
                                    },
                                    "observations": {
                                        "type": "string",
                                        "example": "Paciente presenta leve desviación..."
                                    },
                                    "back_standing_up_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "back_45_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "back_toes_up_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "from_chaplin_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "from_chaplin_toes_up_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "with_insoles_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atención actualizada correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Atención"
                ],
                "summary": "Eliminar una atención",
                "operationId": "61a747753c09bb7fb78bb620c6ef475b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la atención a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminación exitosa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attentions/restore/{id}": {
            "put": {
                "tags": [
                    "Atención"
                ],
                "summary": "Activar un consultorio",
                "operationId": "1790a9e199af22e05554c6afb06b301c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Atención activa"
                    },
                    "404": {
                        "description": "Atención no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación para acceder a la API",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}